0.26
Created utils for global refresh and to manage the network status. Created service files to get info from API and also the external json file. Removed data files since providers already take care of all the logic. Created the online and offline mode switch and created a cache method to cache product data so the application does not re-fetch the data from the API everytime the user re-visits the screen.